minor fixes meali list

+++ database/migrations/2022_10_02_195136_create_reviews_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id');
             $table->foreignId('recipe_id');
-            $table->integer('rating');
+            $table->decimal('rating', 10, 1);
             $table->timestamps();
         });
     }
